requirements.txt
glfw==1.4.0
multipledispatch==0.4.9
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
PyOpenGL==3.1.0
pyrr==0.9.2
